bootloader: Modularize the bootloader code.
Split the bootloader code into multiple files to make the code cleaner.
