+++ src/bootloader/bootloader_hash.c
@@ -1,0 +1,142 @@
+#include "bootloader_hash.h"
+
+#include <driver_init.h>
+#include <platform/platform_config.h>
+#include <pukcc/curve_p256.h>
+#include <util.h>
+#include <utils.h>
+
+COMPILER_ALIGNED(128)
+static struct sha_context _pukcc_sha256_context;
+COMPILER_PACK_RESET()
+
+// clang-format off
+#if PRODUCT_BITBOX_BTCONLY == 1
+static const uint8_t _root_pubkeys[BOOT_NUM_ROOT_SIGNING_KEYS][BOOT_PUBKEY_LEN] = { // order is important
+    {
+        0xe7, 0x30, 0x92, 0x3f, 0x19, 0x04, 0xd3, 0xe8, 0x3e, 0x0a, 0xb0, 0x7f, 0x05, 0x0d, 0x8e, 0x4e,
+        0x60, 0x4c, 0x46, 0xba, 0x36, 0xae, 0xbb, 0xdb, 0x9a, 0x4d, 0x3c, 0x96, 0x60, 0x5b, 0x9e, 0xc3,
+        0x94, 0x71, 0x03, 0x5c, 0x0f, 0x87, 0x34, 0xbd, 0xcb, 0xe1, 0x8f, 0x0c, 0xaa, 0x6f, 0x6e, 0x3d,
+        0x8b, 0x7f, 0xa0, 0x5e, 0x42, 0x5a, 0xc2, 0x21, 0x5e, 0x6e, 0x97, 0x97, 0x5f, 0x48, 0x03, 0x8b,
+    },
+    {
+        0xfd, 0x11, 0xed, 0xc3, 0xb6, 0xd8, 0xe5, 0x0b, 0xe8, 0x6b, 0x2e, 0x33, 0xee, 0xb0, 0x3b, 0x32,
+        0x23, 0x91, 0x4c, 0x12, 0xe4, 0x9d, 0x0e, 0x64, 0xd9, 0x74, 0x03, 0xa7, 0x31, 0xc4, 0x30, 0x6a,
+        0x44, 0xab, 0x6a, 0x99, 0x3e, 0x9d, 0xaa, 0x44, 0xda, 0x0e, 0x5a, 0x0a, 0x8b, 0x3e, 0x02, 0xe9,
+        0xa4, 0x6e, 0x2c, 0xb6, 0x14, 0x57, 0xce, 0x78, 0x81, 0xa8, 0x71, 0x55, 0xee, 0x23, 0x67, 0x44,
+    },
+    {
+        0x75, 0x98, 0x7e, 0x69, 0xa5, 0xed, 0xa5, 0x3f, 0x79, 0x63, 0x16, 0xfa, 0x47, 0x00, 0xf9, 0x9a,
+        0x86, 0x36, 0xb0, 0xa5, 0x6c, 0x57, 0x28, 0xee, 0x8a, 0xd3, 0xb3, 0xcc, 0x8f, 0x37, 0xe6, 0xac,
+        0xfc, 0xa6, 0x08, 0x23, 0x14, 0x4a, 0xea, 0xb2, 0xe4, 0xa7, 0x62, 0x97, 0x89, 0xd0, 0x3e, 0xa4,
+        0xd2, 0xd1, 0x8a, 0xbf, 0x0f, 0xd7, 0xa0, 0x07, 0xd7, 0x96, 0xa1, 0x65, 0x57, 0x28, 0x4b, 0x3f,
+    },
+};
+#elif PRODUCT_BITBOX_MULTI == 1
+static const uint8_t _root_pubkeys[BOOT_NUM_ROOT_SIGNING_KEYS][BOOT_PUBKEY_LEN] = { // order is important
+    {
+        0x2b, 0x8a, 0x21, 0x9f, 0x01, 0x57, 0xf3, 0x12, 0x3e, 0x3c, 0x69, 0xd5, 0x40, 0x0e, 0xff, 0xfe,
+        0xd9, 0x55, 0x3f, 0x64, 0xd2, 0x5e, 0xec, 0x10, 0x65, 0x7d, 0x0e, 0x77, 0x6a, 0x8e, 0x17, 0xe2,
+        0xac, 0x2c, 0x3d, 0x19, 0xca, 0xfd, 0x18, 0x9b, 0x62, 0x36, 0x68, 0xc4, 0x10, 0x1b, 0x9b, 0x96,
+        0x74, 0xf5, 0xe6, 0xe6, 0x43, 0xa3, 0x82, 0x8e, 0x6a, 0xb9, 0x3b, 0x46, 0xe9, 0xc3, 0x87, 0x73,
+    },
+    {
+        0x1e, 0xb6, 0xe7, 0xad, 0x36, 0x1c, 0x40, 0x6b, 0xdd, 0xf4, 0x47, 0x17, 0x7b, 0x38, 0x21, 0x52,
+        0x7a, 0x7a, 0x25, 0x75, 0x99, 0x5c, 0x6f, 0x49, 0xc2, 0xd8, 0xe0, 0x9b, 0x89, 0x7a, 0x70, 0xb7,
+        0xab, 0x9c, 0x53, 0xa4, 0xcf, 0xbe, 0x29, 0x13, 0x33, 0x27, 0x1e, 0xd6, 0x75, 0x6b, 0xd3, 0xad,
+        0x20, 0x77, 0x08, 0x2d, 0xe7, 0x1a, 0xcd, 0xc8, 0xb8, 0x7c, 0x39, 0x2d, 0x94, 0x64, 0xef, 0x51,
+    },
+    {
+        0x9f, 0x39, 0x3d, 0x4a, 0xb6, 0xe8, 0xb0, 0xda, 0xc7, 0xc0, 0x43, 0x62, 0xc5, 0x20, 0xe1, 0xbc,
+        0x1b, 0x32, 0x7d, 0x92, 0xee, 0x25, 0x6b, 0xc4, 0x07, 0xf6, 0x82, 0xff, 0xac, 0x12, 0x0d, 0xec,
+        0x6b, 0x80, 0x23, 0x28, 0x62, 0xb9, 0x7d, 0xfd, 0x3d, 0xcf, 0x21, 0xdb, 0xb3, 0xf5, 0x4e, 0x4c,
+        0x1b, 0xec, 0x2b, 0x99, 0x40, 0x61, 0x3e, 0xef, 0xe7, 0x9e, 0x6a, 0x2b, 0x10, 0xbe, 0xda, 0xc2,
+    },
+};
+#elif PRODUCT_BITBOX_BASE == 1
+static const uint8_t _root_pubkeys[BOOT_NUM_ROOT_SIGNING_KEYS][BOOT_PUBKEY_LEN] = { // order is important
+    {
+        0x89, 0x61, 0xff, 0xaa, 0x89, 0x9b, 0x9d, 0x2d, 0x71, 0xee, 0x31, 0x62, 0xe2, 0x01, 0x20, 0xbe,
+        0x63, 0x0a, 0xfb, 0x5d, 0xf3, 0x5c, 0x99, 0xbf, 0x10, 0xa9, 0x15, 0xbf, 0x48, 0xd2, 0xc6, 0xf6,
+        0xa0, 0x83, 0xbe, 0x8b, 0xb5, 0x35, 0xf7, 0x69, 0x67, 0xc8, 0x2a, 0xa5, 0xdd, 0x5a, 0xe8, 0x7e,
+        0x3e, 0xe6, 0xa0, 0x4d, 0x87, 0x9c, 0xba, 0x08, 0x3a, 0x46, 0x63, 0x62, 0xb7, 0xc1, 0x86, 0x35,
+    },
+    {
+        0x3a, 0xb8, 0x77, 0x41, 0x6e, 0x4e, 0x1d, 0xfc, 0xcb, 0x5c, 0xec, 0xdd, 0xf5, 0xfb, 0x42, 0x18,
+        0x51, 0x1b, 0xe3, 0x5e, 0x23, 0x06, 0x9e, 0x86, 0x20, 0x0c, 0x79, 0x7e, 0xc7, 0xc5, 0xd7, 0xe3,
+        0x04, 0x1e, 0xc7, 0x23, 0xec, 0x1e, 0xae, 0xbe, 0xf2, 0x22, 0xad, 0x60, 0xda, 0x93, 0xea, 0xe7,
+        0xf0, 0x60, 0x48, 0xd1, 0x66, 0x34, 0xf9, 0xf0, 0xcb, 0xa9, 0x7a, 0x26, 0x50, 0x2a, 0x01, 0xd1,
+    },
+    {
+        0x03, 0xc9, 0xb2, 0xd1, 0xea, 0x13, 0x6a, 0x0c, 0x5f, 0xe5, 0x01, 0x91, 0x32, 0xe6, 0x57, 0xf4,
+        0x2d, 0xfa, 0x75, 0x8f, 0x74, 0xb0, 0xb3, 0xe3, 0x74, 0x0e, 0x51, 0xdc, 0x5f, 0x7f, 0xd6, 0x0d,
+        0xdb, 0x00, 0x94, 0x7a, 0xa3, 0x2a, 0xff, 0x05, 0x20, 0xa9, 0x66, 0xfc, 0xe7, 0xdd, 0x2a, 0x81,
+        0x8e, 0x6c, 0xd1, 0x48, 0xc3, 0x54, 0x8a, 0x9d, 0xfb, 0x6b, 0xdf, 0x61, 0xd6, 0xc4, 0xf3, 0xee,
+    },
+};
+#else
+#error "unknown product"
+#endif
+
+static const uint8_t _empty_sig[BOOT_SIG_LEN] = {0};
+
+static void _double_hash(const uint8_t* data, uint32_t len, uint8_t* hash)
+{
+    pukcc_sha256_compute(data, len, hash);
+    pukcc_sha256_compute(hash, SHA256_DIGEST_LENGTH, hash);
+}
+
+void bootloader_hash_signing_keys(const boot_data_t* data, uint8_t* hash_out)
+{
+    _double_hash(
+        (const uint8_t*)(&data->fields.signing_pubkeys_version),
+        sizeof(data->fields.signing_pubkeys_version) + sizeof(data->fields.signing_pubkeys),
+        hash_out);
+}
+
+// double hashes firmware version | firmware
+void bootloader_hash_firmware(const boot_data_t* data, uint8_t* hash_out)
+{
+    sha_sync_sha256_start(&HASH_ALGORITHM_0, &_pukcc_sha256_context, false);
+    sha_sync_sha256_update(
+        &HASH_ALGORITHM_0,
+        (const uint8_t*)&data->fields.firmware_version,
+        sizeof(data->fields.firmware_version));
+    sha_sync_sha256_update(&HASH_ALGORITHM_0, (const uint8_t*)FLASH_APP_START, FLASH_APP_LEN);
+    sha_sync_sha256_finish(&HASH_ALGORITHM_0, hash_out);
+    pukcc_sha256_compute(hash_out, SHA256_DIGEST_LENGTH, hash_out);
+}
+
+bool bootloader_hash_is_empty_sig(const uint8_t* sig) {
+    return MEMEQ(sig, _empty_sig, sizeof(_empty_sig));
+}
+
+secbool_u32 bootloader_hash_pubkeys_verified(const boot_data_t* data)
+{
+    // Verify the signing keys' version and pubkeys, signed by the root keys
+    volatile uint8_t valid = 0;
+
+    uint8_t hash[SHA256_DIGEST_LENGTH];
+    bootloader_hash_signing_keys(data, hash);
+    for (uint8_t i = 0; i < BOOT_NUM_ROOT_SIGNING_KEYS; i++) {
+        const uint8_t* sig =
+            (const uint8_t*)&data->fields.root_signatures_of_signing_pubkeys + i * BOOT_SIG_LEN;
+        uint8_t random[SHA256_DIGEST_LENGTH];
+        if (bootloader_hash_is_empty_sig(sig)) {
+            continue;
+        }
+        rand_sync_read_buf8(&RAND_0, random, sizeof(random));
+        // First is sanity check
+        if (pukcc_ecdsa_verify(_root_pubkeys[i], sig, random, sizeof(random), curve_p256) != 0) {
+            if (pukcc_ecdsa_verify(_root_pubkeys[i], sig, hash, sizeof(hash), curve_p256) == 0) {
+                valid++;
+                continue;
+            }
+        }
+        return secfalse_u32;
+    }
+    if (valid >= BOOT_ROOT_SIG_M) {
+        return sectrue_u32;
+    }
+    return secfalse_u32;
+}
